build: simplify
