+++ src/cli/index.ts
@@ -1,5 +1,20 @@
 #!/usr/bin/env node
 
-import { main } from './main.ts';
+import { createRequire } from 'node:module';
+import { isSea } from 'node:sea';
 
-void main();
+if (isSea()) {
+  const __require = createRequire(import.meta.url);
+
+  Object.assign(globalThis, {
+    __bundlerPathsOverrides: {
+      'thread-stream-worker': __require.resolve('./thread-stream-worker.js'),
+      'pino/file': './pino-file.js',
+      'pino-pretty': './pino-pretty.js',
+      'pino-worker': './pino-worker.js',
+    },
+  });
+}
+
+const m = await import('./main.ts');
+await m.main();
